backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
gunicorn==22.0.0
python-dotenv==1.0.1
pydantic==2.10.4
pydantic-settings==2.7.1
sqlalchemy==2.0.36
aiosqlite==0.20.0
bcrypt==4.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
langchain==0.3.14
langchain-community==0.3.14
langchain-openai==0.3.0
langchain-google-genai>=2.0.0
langchain-chroma==0.2.2
chromadb==0.5.23
openai==1.58.1
tiktoken==0.8.0
sentence-transformers==3.3.1
httpx==0.28.1
google-generativeai>=0.8.0
